Buffer pen lines
This changes pen rendering to be buffered so that multiple lines will be
drawn in a single draw call, instead of 1 draw call per line. This can
result in significant performance improvements.

This implementation is a hack, definitely doesn't belong in upstream.
The performance changes are likely not noticeable in Scratch 3 because
scratch-vm is the bottleneck. In my scratch-vm fork, however, the
performance difference is roughly ~4x faster on my computer. There is
still lots of room for improvement.

There's definitely some weird edge cases I missed, and it definitely
will misbehave when running on a mobile phone with a GPU from the stone
age.

This might be slower in some edge cases where buffering is just extra
overhead.